+++ release-notes.md
@@ -1,4 +1,16 @@
 # Orion 15.13.0 #
+
+## Enabling auto archive when using live with recording ##
+- Issue Type: Task
+- Issue ID : WEBC-1574
+
+### Configuration ##
+none.
+
+#### Deployment Scripts ####
+Run 'php /opt/kaltura/app/deployment/updates/scripts/2019_12_10_update_archive_permissions.php'
+
+
 ## Reach boolean event notification for privacy context ##
 - Issue Type: Task
 - Issue ID : REACH2-737